+++ packages/playground-ui/src/utils.ts
@@ -1,0 +1,2 @@
+// Utility functions that can be imported without React dependencies
+export { isWorkspaceV1Supported } from './domains/workspace/compatibility';
