+++ packages/playground-ui/vite.config.ts
@@ -30,6 +30,7 @@ export default defineConfig({
     lib: {
       entry: {
         index: resolve(__dirname, 'src/index.ts'),
+        utils: resolve(__dirname, 'src/utils.ts'),
         tokens: resolve(__dirname, 'src/ds/tokens/index.ts'),
         'tailwind.preset': resolve(__dirname, 'tailwind.preset.ts'),
       },
